Added KeymasterUI component
